进一步优化DotNet.Util.Db 其它一些方法重构和重命名
