+++ src/DotNet.Business/DotNet.Business.csproj
@@ -5,7 +5,7 @@
 		<DisableImplicitNamespaceImports>true</DisableImplicitNamespaceImports>
 		<Description>DotNet.Business is sample for User,Role,Menu,Module,Permission etc Management. A set of .NET 5/6/Standard 2.0/Standard 2.1 utilities(partial but great majority support .net framework 4.0/4.5/4.6). Please visit https://github.com/cuiwenyuan/DotNet.Util for more information</Description>
 		<Company>www.cuiwenyuan.com</Company>
-		<Version>1.0.8</Version>
+		<Version>1.0.9</Version>
 		<Authors>崔文远Troy Cui</Authors>
 		<PackageReleaseNotes>Please visit https://github.com/cuiwenyuan/DotNet.Util for details</PackageReleaseNotes>
 		<PackageTags>DotNet Business</PackageTags>
@@ -22,8 +22,8 @@
 		<SymbolPackageFormat>snupkg</SymbolPackageFormat>
 		<ApplicationIcon>App.ico</ApplicationIcon>
 		<RepositoryType>git</RepositoryType>
-		<AssemblyVersion>5.6.2022.0511</AssemblyVersion>
-		<FileVersion>5.6.2022.0511</FileVersion>
+		<AssemblyVersion>5.6.2022.0514</AssemblyVersion>
+		<FileVersion>5.6.2022.0514</FileVersion>
 		<LangVersion>latest</LangVersion>
 	</PropertyGroup>
