+++ src/DotNet.Util/DotNet.Util.csproj
@@ -4,7 +4,7 @@
 		<TargetFrameworks>net40;net45;net46;net47;net48;net5.0;net6.0;netstandard2.0;netstandard2.1</TargetFrameworks>
 		<Description>A set of .NET 5/6/Standard 2.0/Standard 2.1 utilities(partial but great majority support .net framework 4.0/4.5/4.6/4.7/4.8). Please visit https://github.com/cuiwenyuan/DotNet.Util for more information</Description>
 		<Company>www.cuiwenyuan.com</Company>
-		<Version>1.0.29</Version>
+		<Version>1.0.30</Version>
 		<Authors>崔文远Troy Cui</Authors>
 		<PackageReleaseNotes>Please visit https://github.com/cuiwenyuan/DotNet.Util for details</PackageReleaseNotes>
 		<PackageTags>DotNet Util Utilities</PackageTags>
@@ -21,8 +21,8 @@
 		<SymbolPackageFormat>snupkg</SymbolPackageFormat>
 		<ApplicationIcon>App.ico</ApplicationIcon>
 		<RepositoryType>git</RepositoryType>
-		<AssemblyVersion>5.6.2022.0511</AssemblyVersion>
-		<FileVersion>5.6.2022.0511</FileVersion>
+		<AssemblyVersion>5.6.2022.0514</AssemblyVersion>
+		<FileVersion>5.6.2022.0514</FileVersion>
 		<LangVersion>latest</LangVersion>
 	</PropertyGroup>
